Break GB classes out into individual files
