+++ lib/greenbutton/models.rb
@@ -1,0 +1,9 @@
+require 'greenbutton/models/green_button_entry'
+require 'greenbutton/models/electric_power_usage_summary'
+require 'greenbutton/models/electric_power_quality_summary'
+require 'greenbutton/models/interval_block'
+require 'greenbutton/models/interval_reading'
+require 'greenbutton/models/local_time_parameters'
+require 'greenbutton/models/meter_reading'
+require 'greenbutton/models/reading_type'
+require 'greenbutton/models/usage_point'
